Analysis using the new data.
A.0.0. The initial analysis contains plots for measures of the entire
data set. This was done to investigate if anything stands out.
The enviroments still look cooperative.

A.1. Is the analysis at the tournament level of different types.
Now we are ploting the transitions to C.

Everything is still in primary stage.

Last notebook contains the note to myself for
getting back.

The notes.tex contains some notes for the variables.