Make copy edits and add some TODO tags.
One of the TODO tags is about a particular paragraph that I really
struggle with. If I'm alone in not liking the paragraph then I'm happy
for it to be left as it is (I'm not
entirely sure I could offer my own version of it as I'm not sure I
understand it).