Remove paragraph about TfT
@nikoleta-v3 did some basic analysis (she's going to put the plots up on
the PR) and found that:

- TfT's median C/C_mean was 1.2
- Tf2T's median C/C_mean was 1.37
- Random's median C/C_mean was 0.8

So it's not immediately clear (from the data) that TfT does match the
C_mean (which makes sense if it is never provoked) and in fact, on
average is "as far" from that mean as Random is.

I've added a TODO to include some sort of similar discussion in the
final section.